🚚 move non-scenario files outside of the scenario folder
