+++ package.json
@@ -9,8 +9,8 @@
     "build": "lerna run build --stream",
     "build:bundle": "lerna run build:bundle --stream",
     "format": "prettier --check \"**/*.{ts,js,json,md,yml,html}\"",
-    "lint": "scripts/cli lint . && scripts/cli lint test/e2e/scenario",
-    "typecheck": "scripts/cli typecheck . && scripts/cli typecheck test/e2e/scenario",
+    "lint": "scripts/cli lint . && scripts/cli lint test/e2e",
+    "typecheck": "scripts/cli typecheck . && scripts/cli typecheck test/e2e",
     "dev": "node scripts/dev-server.js",
     "release": "lerna version --exact",
     "version": "node ./scripts/generate-changelog.js",
