+++ test/e2e/lib/globalTypes.ts
