+++ test/e2e/ts-node.js
@@ -1,4 +1,4 @@
 module.exports = require('ts-node').register({
   files: true,
-  project: 'test/e2e/scenario/tsconfig.json',
+  project: 'test/e2e/tsconfig.json',
 })
